#12 feat : 미비한 로직들 처리 완료 - HRM - FCM - SD
이를통해 gateway와 완벽 호환이 됩니다.

+++ src/main/java/org/ever/_4ever_be_business/hr/dto/response/StatisticValueDto.java
@@ -12,6 +12,11 @@ public class StatisticValueDto {
     @JsonProperty("value")
     private Integer value;
 
+    /**
+     * Absolute difference from previous period (not percentage)
+     * Example: if current = 15, previous = 12, delta_rate = 3
+     * Example: if current = 10, previous = 12, delta_rate = -2
+     */
     @JsonProperty("delta_rate")
-    private Double deltaRate;
+    private Integer deltaRate;
 }
